# backend/src/main/java/com/app/bookshop/dto/CustomerRequest.java
package com.app.bookshop.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record CustomerRequest(
    @NotBlank @Size(max = 255) String firstName,
    @NotBlank @Size(max = 255) String lastName,
    @NotBlank @Email @Size(max = 255) String email
) {
}