# backend/src/main/java/com/app/bookshop/dto/OrderItemRequest.java
package com.app.bookshop.dto;

import jakarta.validation.constraints.Max;
import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;

public record OrderItemRequest(
    @NotNull Long productId,
    @Min(1) @Max(99) int quantity
) {
}